Add files via upload
The culminating endeavor of the Google Data Analytics certificate I have recently completed is the Cyclistic bike-share capstone project. This undertaking allowed me to effectively employ the array of tools I gained proficiency in throughout the duration of the program, which encompassed R, Tableau, and PowerPoint. The project centered on leveraging these skills to address a practical real-world scenario. Rstudio was used for the data processing phase of this project.